Add specs for container, column, and block Remove old layout files
