+++ src/components/block/index.js
@@ -1,6 +1,3 @@
 import component from "./component"
-// import fixtures from "./fixtures"
 
 export default component
-
-// export { component, fixtures }
