+++ src/components/container/index.js
@@ -1,6 +1,6 @@
 import component from "./component"
-// import fixtures from "./fixtures"
+import fixtures from "./fixtures"
 
 export default component
 
-// export { component, fixtures }
+export { component, fixtures }
