--- src/layout/container/component-map.js
@@ -1,13 +1,0 @@
-import Button from "@src/components/button"
-import Content from "@src/components/content"
-import Form from "@src/components/form"
-import Frame from "@src/components/frame"
-import Image from "@src/components/image"
-
-export default {
-  "component-button": Button,
-  "component-content": Content,
-  "component-form": Form,
-  "component-frame": Frame,
-  "component-image": Image
-}
