--- src/layout/container/index.js
@@ -1,8 +1,0 @@
-import component from "./component"
-import componentMap from "./component-map"
-import container from "./container"
-import fixtures from "./fixtures"
-
-export default container
-
-export { component, componentMap, container, fixtures }
